extract modules
